+++ Sources/TuistEnvKit/Commands/InstallCommand.swift
@@ -17,7 +17,7 @@ struct InstallCommand: ParsableCommand {
         name: .shortAndLong,
         help: "Re-installs the version compiling it from the source"
     )
-    var force: Bool
+    var force: Bool = false
 
     func run() throws {
         try InstallService().run(version: version,