+++ Sources/TuistKit/Commands/GenerateCommand.swift
@@ -10,21 +10,30 @@ struct GenerateCommand: ParsableCommand {
 
     @Option(
         name: .shortAndLong,
-        help: "The path where the project will be generated."
+        help: "The path where the project will be generated.",
+        completion: .directory
     )
     var path: String?
 
     @Flag(
         help: "Only generate the local project (without generating its dependencies)."
     )
-    var projectOnly: Bool
+    var projectOnly: Bool = false
 
     @Flag(help: "Generate a project replacing dependencies with pre-compiled assets.")
-    var cache: Bool
+    var cache: Bool = false
+
+    @Option(
+        name: NameSpecification([.customShort("i"), .customLong("include-sources", withSingleDash: false)]),
+        parsing: .singleValue,
+        help: "When used with --cache, it generates the given target (with the sources) even if it exists in the cache."
+    )
+    var includeSources: [String] = []
 
     func run() throws {
         try GenerateService().run(path: path,
                                   projectOnly: projectOnly,
-                                  cache: cache)
+                                  cache: cache,
+                                  cacheSources: Set(includeSources))
     }
 }