+++ Sources/TuistKit/Commands/TuistCommand.swift
@@ -25,9 +25,16 @@ public struct TuistCommand: ParsableCommand {
                                  CloudCommand.self,
                                  CacheCommand.self,
                                  SigningCommand.self,
+                                 MigrationCommand.self,
                                  CleanCommand.self,
                              ])
     }
+
+    @Flag(
+        name: [.customLong("help-env")],
+        help: "Display subcommands to manage the environment tuist versions."
+    )
+    var isTuistEnvHelp: Bool = false
 
     public static func main(_ arguments: [String]? = nil) -> Never {
         let errorHandler = ErrorHandler()