+++ Sources/TuistKit/Services/GenerateService.swift
@@ -4,12 +4,12 @@ import TuistLoader
 import TuistSupport
 
 protocol GenerateServiceProjectGeneratorFactorying {
-    func generator(cache: Bool) -> ProjectGenerating
+    func generator(cache: Bool, includeSources: Set<String>) -> ProjectGenerating
 }
 
 final class GenerateServiceProjectGeneratorFactory: GenerateServiceProjectGeneratorFactorying {
-    func generator(cache: Bool) -> ProjectGenerating {
-        ProjectGenerator(graphMapperProvider: GraphMapperProvider(cache: cache))
+    func generator(cache: Bool, includeSources: Set<String>) -> ProjectGenerating {
+        ProjectGenerator(graphMapperProvider: GraphMapperProvider(cache: cache, includeSources: includeSources))
     }
 }
 
@@ -26,14 +26,14 @@ final class GenerateService {
         self.projectGeneratorFactory = projectGeneratorFactory
     }
 
-    func run(
-        path: String?,
-        projectOnly: Bool,
-        cache: Bool
-    ) throws {
+    func run(path: String?,
+             projectOnly: Bool,
+             cache: Bool,
+             cacheSources: Set<String>) throws
+    {
         let timer = clock.startTimer()
         let path = self.path(path)
-        let generator = projectGeneratorFactory.generator(cache: cache)
+        let generator = projectGeneratorFactory.generator(cache: cache, includeSources: cacheSources)
 
         _ = try generator.generate(path: path, projectOnly: projectOnly)