+++ Sources/TuistSupport/Constants.swift
@@ -29,6 +29,7 @@ public struct Constants {
     }
 
     public struct EnvironmentVariables {
+        public static let verbose = "TUIST_VERBOSE"
         public static let colouredOutput = "TUIST_COLOURED_OUTPUT"
         public static let versionsDirectory = "TUIST_VERSIONS_DIRECTORY"
         public static let cacheDirectory = "TUIST_CACHE_DIRECTORY"