+++ Sources/tuist/main.swift
@@ -1,8 +1,9 @@
 import Foundation
 import enum TSCBasic.ProcessEnv
+import TuistSupport
 import enum TuistSupport.LogOutput
 
-if CommandLine.arguments.contains("--verbose") { try? ProcessEnv.setVar("TUIST_VERBOSE", value: "true") }
+if CommandLine.arguments.contains("--verbose") { try? ProcessEnv.setVar(Constants.EnvironmentVariables.verbose, value: "true") }
 
 LogOutput.bootstrap()