+++ stats/next.config.js
@@ -1,0 +1,3 @@
+// next.config.js
+const withMDX = require('@next/mdx')()
+module.exports = withMDX()